Move the System Map to PiGUI
The main idea of the interaction of lua side and C++ side is the same as
in the worldview - the C++ View Class builds three-dimensional objects,
some projected points are passed to the lua script, the lua script only works
with screen coordinates.

The difference between the system view and the world view is that not all objects can
have a physical body. If we examine a non-current star system, we see only system
bodies that do not have physical bodies. Therefore, if for a world view you can
simply create an array of physical bodies, this will not work in a system view.

To solve this problem, struct "Projectable" was created, into which you can write
any point object, system body, ship, Lagrange point, apocenter, pericenter, etc.
Now all these objects can be processed in a single array.

The names of the types of these objects are recorded in an enum, which is
exported to a enum_table.cpp and read by the script as integers at the start.
I thought it was better to avoid passing strings in such a hot place in the code.
(runs for every object in every frame) Therefore, type names are written and
transmitted as integers.

All colors are now imported from lua: color index names are listed in the enum,
which is exported to the enum_table.cpp. At event "onGameStart" the lua script
reads these names and sends color values to the class. Now you can change
colors at runtime.

 - Improvements/fixes:

Fix zooming speeds.
Fix ship's orbits in nonroot frame.
Fix system bodies orbits, ..apsis and lagrange points were displayed in the wrong place.
Remove unused old gui objects.
Now only those orbits whose radius is less than 1% of the screen width are not shown.
Landed ship are now spinning with the planet when rewinding time in a planner.
Now you can center the view on any object, including the player’s ship.
Add right-click context menu to center or set object as target. Therefore, to rotate the view, set the middle button.
Add smooth transision when click object.
Add indicators for navigation target and combat target.

 * Changes in the files explained *

 - SystemView.cpp, SystemView.h, system-view-ui.lua, lang/core/en.json

The whole thing, lots of changes. Now the system view draws only the orbits and the grid.
In those places where the icons were drawn, instead of drawing, they are stored in an
array, which is then requested from the lua.

 - LuaEngine.cpp

Lots of static functions to interact between the Lua and C++ side. (sorry)

 - Space.cpp

Added a notification for the systemview when the ship is removed from space, because it can be centered at this time, and an segmentation fault will occur.
Added static_cast, because it caused a warning at compilation.

 - LuaSystemBody.cpp

Added attributes to determine the shape of the icon and set the navigation target.

 - enum_table.cpp, enum_table.h, pigui/Face.h

Rescanned enums to pass color settings and object types between Lua and C++. After scanning an error was found in the pigui/Face.h, fixed.
fixes #4801

 - LuaGame.cpp

The float does not have enough accuracy to display time up to seconds, change to double.

 - icons.svg, themes/default.lua

Add icons for transfer planner, lagrange points, apoapsis, periapsis.