On-screen buttons for rotating and zooming view
Instead of the separate zoom in and zoom out buttons, the zoom
and rotation button is now. You need to click on the button and move the mouse,
and then the view will zoom or rotate accordingly.
Add a special flag (doingViewTransformation) to let the game know when
mouse events should pass through the windows of the Imgui. When this flag
is active, mouse capture is activated, and mouse events are also transmitted to Gui and
input. Add functions to manage this flag.

Fix icon name for lagrange marker.