Add target follow mode (draft)
2 fixspeed modes - FLYING and DOCKING
Automatically switch when lowering / raising the chassis.
Differ in which axis the flight speed is fixed.
For FLYING the axis is -Z (forward).
For DOCKING the axis iz Y (up).
When you press W/S, R/F, respectively, setspeed is set to the value with
which the ship is actually moving, depenging on the fixspeed mode
( FLYING / DOCKING )

Add FOLLOW_IT item to radial menu for ships and stations.
When activating the follow mode, an additional button appears to switch
the follow mode.
There is a POS follow mode, when the ship maintains a position relative
to the target, in the coordinate system of the main frame.
And also the ORIENT mode in which the ship maintains the relative
position and orientation in the object's coordinate system.

add speed limiter