DEBUG: Add a "debug pause"
I called it that because it is different from a normal pause. Objects in
the game can react to changes in time acceleration, unfortunately this
seems inevitable. The debug pause is invisible to anyone, time
acceleration remains. It activates via <c-0>. And most importantly, when
pressed again, exactly 1 timestep occurs with the specified time
acceleration. Deactivated by <c-s-0>.

Also added a tab to the Debug Info window with a slider that allows to
change the frame interpolation coefficient and a “Next Frame” button.