Remove ports from space stations
The port was apparently conceptually conceived as an almost isolated
group of landing pads, with a common entrance and exit, as well as
identical dimensions.

At the moment it was used to store the sizes of the corresponding group
of pads, and to lock this group if a landing occurs on one of them.

Each pad's dimensions will now be stored within the pad itself, and
locking will be temporarily disabled. Anyway, at the orbital station we
have 6 ports superimposed on each other in one entrance. In the future,
it is planned to lock docking based on capturing waypoints, that will be
reused by the pads.

For compatibility with existing docking tags, the code for loading these
tags from the model is left.