Show uniform percentage in hyperspace
Long ago, hyperjump progress was shown as a "probability of arrival",
which increased unevenly, slowing down to 100%. This was then renamed to
simply "jump complete", but the unevenness remained. Apparently the
concept of hyperjump itself has changed since then. I think the uniform
flow of hyperjump progress is now more consistent both with the name in
the GUI and with the animation in the sector map.